+++ app/routes/transactions.js
@@ -37,8 +37,9 @@ export default Ember.Route.extend({
       this.transitionTo('transactions.edit', entity);
     },
     delete: function (entity) {
-      entity.destroyRecord();
-      this.get("controller").set('isModalDialogActive', false);
+      if (confirm('Are you sure')) {
+        entity.destroyRecord();
+      }
     }
   }
 });
